refactor item API
