+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.core.validation.ValidId;
+import ru.practicum.shareit.core.validation.validid.ValidId;
 import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.dto.UpdateItemDto;
 import ru.practicum.shareit.item.service.ItemService;
@@ -23,7 +23,10 @@ public class ItemController {
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
-    public List<ItemDto> getAllByOwner(@RequestHeader(value = X_SHARER_USER_ID, required = false) @ValidId(message = INVALID_USER_ID_MESSAGE) Long userId) {
+    public List<ItemDto> getAllByOwner(
+            @RequestHeader(value = X_SHARER_USER_ID, required = false)
+            @ValidId(message = INVALID_USER_ID_MESSAGE) Long userId
+    ) {
         return itemService.getAllByOwner(userId);
     }
