+++ src/main/java/ru/practicum/shareit/request/mapper/ItemRequestMapper.java
@@ -2,6 +2,7 @@
 package ru.practicum.shareit.request.mapper;
 
 import lombok.experimental.UtilityClass;
+import ru.practicum.shareit.request.dto.CreateItemRequestDto;
 import ru.practicum.shareit.request.dto.ItemRequestDto;
 import ru.practicum.shareit.request.model.ItemRequest;
 import ru.practicum.shareit.user.model.User;
@@ -21,7 +22,7 @@ public class ItemRequestMapper {
         );
     }
 
-    public ItemRequest toItemRequest(ItemRequestDto itemRequestDto, User requestor) {
+    public ItemRequest toItemRequest(CreateItemRequestDto itemRequestDto, User requestor) {
         ItemRequest itemRequest = new ItemRequest();
         itemRequest.setDescription(itemRequestDto.getDescription());
         itemRequest.setRequestor(requestor);
