+++ src/main/resources/schema.sql
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS items (
   description VARCHAR(512) NOT NULL,
   is_available BOOLEAN NOT NULL,
   owner_id BIGINT NOT NULL,
-  request_id BIGINT NOT NULL,
+  request_id BIGINT,
   CONSTRAINT pk_item PRIMARY KEY (id),
   CONSTRAINT fk_item_owner_id FOREIGN KEY (owner_id) REFERENCES users (id),
   CONSTRAINT fk_item_request_id FOREIGN KEY (request_id) REFERENCES requests (id)
